# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Surplus API Documentation",
        "description": "L5 Swagger OpenApi Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/images": {
            "get": {
                "tags": [
                    "images"
                ],
                "summary": "get list of images",
                "description": "images list",
                "operationId": "f2dc6a8cdd6300fbb2b54a95d9061b77",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "images"
                ],
                "summary": "create images",
                "description": "create images",
                "operationId": "createimage",
                "requestBody": {
                    "description": "create images",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "file",
                                        "format": "file",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully Created"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    }
                }
            }
        },
        "/images/{id}": {
            "get": {
                "tags": [
                    "images"
                ],
                "summary": "get image by id",
                "description": "get image by id",
                "operationId": "getimagebyid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "image id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "images"
                ],
                "summary": "update image by id",
                "description": "update image by id",
                "operationId": "updateimage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "image id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "update image by id",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "file",
                                        "format": "file",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unprocessable Content - Validation"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "images"
                ],
                "summary": "delete image by id",
                "description": "delete image by Id",
                "operationId": "deleteimage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete image by id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success - No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    }
}